% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{print.punycoder_parsed_url}
\alias{print.punycoder_parsed_url}
\title{Print method for punycoder parsed URL results}
\usage{
\method{print}{punycoder_parsed_url}(x, ...)
}
\arguments{
\item{x}{A punycoder_parsed_url object}

\item{...}{Additional arguments (ignored)}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print method for punycoder parsed URL results
}
