% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/punycoder.R
\docType{package}
\name{punycoder-package}
\alias{punycoder}
\alias{punycoder-package}
\title{Unicode and Punycode Domain Name Processing}
\description{
Provides high-performance functions for encoding and decoding
internationalized domain names according to RFC 3492 (Punycode)
and IDNA standards.
}
\details{
The punycoder package fills a critical gap in R's ecosystem for
handling international domain names. It provides reliable, fast
conversion between Unicode and ASCII representations of domain names.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bart-turczynski/punycoder}
  \item \url{https://CRAN.R-project.org/package=punycoder}
  \item Report bugs at \url{https://github.com/bart-turczynski/punycoder/issues}
}

}
\author{
\strong{Maintainer}: Bart Turczynski \email{bartek+punycoder@turczynski.pl}

Authors:
\itemize{
  \item Bart Turczynski \email{bartek+punycoder@turczynski.pl}
}

}
